requirements.txt
Pillow 
torch
transformers